Implement framework for project-level permissions
This adds a base class (maybe to be merged with ProjectSpecific later)
which adds a way to check permissions by object. What we are aiming for,
however, is not object-level permissions, but project-level. The new
object-level method `attempt` takes an action (i.e. change/delete) and a
user as an argument, and looks for the method 'allow_[action]_for',
which tests a user's project permissions against the object's
affiliation. So, instead of granting a user (or a group) object-level
permissions over every single object with the same project affiliation,
this can accomplish the same with many fewer (project-level)
permissions.

The default `allow_delete_for` method checks if any projects
unaffiliated with a user deleting an object are using that object, and
if so, does not allow the delete to go through (unless the deleter is a
superuser)