Update document admin view
1. Use ZoteroWidget instead of javascript template rendering. I listed
some motivations for this in a previous commit, but an additional reason
is that an ajax call to the server is already necessary for the JS
template rendering function-- so why not leave out the extra step and
render the form on the server, where it can be better controlled?

2. Add support for editing scans & external links in the admin view. The
changes in the base admin template were included in the commit authored
just before this one.

Todo still: link to transcript; related topics.