Data migration: note sections back to citations
It might be worth defining note sections as their own entitites again in
the future, but it was too complicated to come up with a common
interface for editing note sections & citations when they were separate
models (that happened to really do the same thing). If we define more
types of note sections (timelines, text blocks, queries), then we can
make those into their own models, and have note section be a generic
foreign key to them.