Update footnote formset when transcript is edited.
If footnotes are deleted, remove them from the formset by hiding them
and changing the value of the 'DELETE' input checkbox to true. If
footnotes are restored (through 'undo' on the composer), do the reverse.

This checks the count of footnotes in the transcript on every keypress,
cut, paste, undo, and redo. It doesn't seem to slow things down on my
machine, and I don't know if there's any smarter way to do things.

I wish wysihtml5 would forward key{press,up,down} events.