Allow users to log out regardless of auth backend
Logging out of BrowserID requires calling navigator.id.logout()
client-side, meaning that the "logout" link does a preventDefault to
stop going to the logout page until that function finishes. Problem is,
that redirection to the logout page only occurs if a user is, in fact,
authenticated with BrowserID. We're using the regular django auth
backend as well right now, so it's not necessarily true that a user will
be logged in with BrowserID.

This adds a variable specifying whether a user is logged in with the
BrowserID auth backend, making the logout work normally if not.