Be nice to non-ascii in making slugs (breaks urls)
In a new slug fn, instead of finding the closest ascii replacement to a
non-ascii character, we accept all unicode alphanumeric characters, as
well as the following characters: '._-,' (not the quote marks).
Additionaly, we do not convert to lower case, and words are delineated
with underscores instead of hyphens.

The previously slugging function didn't work with non-latinish alphabets
such as Cyrillic which can't be converted to ascii by just removing
diacritics, etc.

This breaks basically every topic URL, but if there's a time to do that,
it's now.

p.s. did you know that re.sub doesn't take flags? and did you know that
re.UNICODE equals 32? and did you know that if you provide re.UNICODE
naively as a flag for the fourth argument to re.sub, what you are
actually saying is re.sub(blah, blah, blah, count=32)? well, I know that
now.