New "distributed" topic model(s)
Topics are now recast as "topic nodes" that contain connections to
projects. For now, those connections are names, summaries, and topic
assignments. In the future it's easy to imagine structured data will
fall in the same category.

Also added fixtures for basic multi-project & user setup. The tests for
the new topics rely on these fixtures.