Optimize & add to  note serializer
This was not exactly premature-- there were a massive number of joins on
each query because we weren't selecting related documents/notes to note
sections. This should make that constant.

Maybe instead of doing that here, it should be done with custom managers
for each type of note section. I dunno.