Get rid of tabs on note template
This changes the note template to be a single page. Sections will look
the same, but the "about" and "description" portions have been
significantly modified.

The following metadata is included in a well header:
    * Status
    * Related topics
    * Project
    * Author(s)
    * License
    * Last updated timestamp

The "description" of the note is marked off in its own box so as to
appear somewhat distinct from the note sections (citations, text,
references to other notes, etc.) that follow. There is still some work
to be done on changing the design of this box, but especially the note
sections. They are a bit too wide to be comfortably read, probably.

Additionally, long notes (e.g. with a lot of citations) can be a bit
long & unwieldy. Is that okay? Should those be paginated? Should there
be a type of section just meant for categorization/anchoring? Things to
consider.

Still todo, besides that rethinking just mentioned:
    1. Affordances for editing notes
    2. Update backbone app to work with new template
    3. Display of assigned users (should that be for everyone to see?)