Make sure ES responses preserve document order
This should be fine?

The place where the response was loaded was in
pyelasticsearch.client.ElasticSearch._decode_response, which in turn
called requests.models.Response.json. Both of those have a lot of
compatibility related code which I have just completely foregone. We're
assuming python 2.7 already though, so this *shouldn't* be an issue
until an upgrade to python 3 (?) which would be a spicy meatball in
general (duh)