Tweak API filter for ElasticSearch searches
Still need to work on the proper mapping & searching techniques, but
this is closer. Makes all ES models have a display_title field which can
be relied to be searched on (for title matching). This is a multi-field
with a string type and an autocomplete type. I still haven't tamed the
latter so I'm not using it anywhere (yet).