Add new ElasticSearch activity index
The index is based off of reversion Version objects and is updated every
time a new Revision is made. In making the new index, I made a base
class for both it and the item index which includes common methods to
both.

Also deleted an empty file (editorsnotes/search/query.py)