View for editing project roles
This is a little hacky-- I'm doing in javascript what I should be doing
in Django: iterating through the checkboxes in the formset and
constructing a form with swapped x and y axes. This is necessary because
support for iterating through a ModelMultipleChoiceField was only added
in Django 1.6. Once we upgrade, we should change that. Until then, this
is ok, though.