Add unique constraint for Document.description
This can't be done directly at the database level, since description is
an XML field, and such fields don't have equality operators in Postgres.
Instead, we create a md5 digest of the text value of that field, with
all tags, html entities, and all characters not alphanumeric or
whitespace stripped. This field then has a unique constraint in the
database along with the `project` field.