Mixin for project-specific item serializers
Before, we set an object's project in the `pre_save` method of a view.
That was a problem, because `full_clean` is called before `pre_save`,
meaning any cleaning/validation methods of an object which require
access to the project (i.e. unique constraints that are
project-specific) could not accurately be called.

Now, the serializer for any project-specific item should inherit the
ProjectSpecificItemMixin. This first ensures that the serializer is
passed a context containing a project, and then adds that project to the
constructed instance in the `restore_object` method. The project
attribute is then set when the restored object's `full_clean` method is
called.