Make revisions in API views with class decorator
Previously, we had been overriding serializers' `save_object` methods in
order to make revisions. This was never a good way to go about this for
a couple reasons. First, it required that a view pass a context variable
to the serializer indicating that a revision should be saved. Both the
view and the serializer, then, would have to know that the serializer
contained that overriden `save_object` method, which was error-prone,
redundant, and annoying. Second, that's really not what
serialization/deserialization is for. Creating revisions should be an
action initiated with views, not serializers (or forms).

This commit introduces a class decorator called
`@create_revision_on_methods`, which is pretty self-explanatory. Classes
wrapped with this decorator will create revisions when calling any of
the methods provided as arguments. So, List API views (which are POSTed
to), are decorated with `@create_revision_on_methods('create')`,
catching the `rest_framework.mixins.CreateModelMixin.create` method
inherited by that class.

It seems like this decorator could be applied to any class which
contains methods which create/change objects (ie other CBVs), so it
could prove to be useful in other places too.

(all API tests which test revision creation pass now)