Change jquery shim to work more better
In 03eb167, we started to shim non-commonjs modules. Jquery plugins are
terrible, however, and require window.jQuery. In the interest of
simplicity, I've added a shimmed jquery (which is exposed to `window`)
in editorsnotes_app/js/jquery.js which includes all the plugins that we
use. Instead of doing require('jquery'), modules should require that
file.

Also, disabled jquery.ba-bbq, because it's not compatible with
jquery@1.11. I think it's because of browser detection code being
removed. It's ok, because we shouldn't even really need it anyway with
Backbone able to take care of routing.