Create log activity model
This commit changes the way we log user activities. Previously, newly
created reversion.models.Version instances would trigger updates to the
elasticsearch activity index.

There is now an explicit LogAction model which is basically the same as
django.contrib.admin.models.LogEntry. New instances of that are created
in API views in the generic post_save and post_delete methods, which
trigger identical additions to the elasticsearch activity index as
before.

This was necessary in the first place because the newest version of
django reversion does not differentiate delete actions, and we want to
log delete actions. Now that we can update reversion, we can update to
Django 1.6 (I think?)

(I didn't want to use the Django LogEntry model (or inherit it) because
we don't import that module. That prevents us from pulling in the
associated i18n files which are fairly important everywhere in Django.)

unrelated: The Giants just won the pennant on a Travis Ishikawa walkoff
homerun